i#8076: comment change and removed unnecessary typecast

+++ core/kernel_linux/os.c
@@ -108,8 +108,8 @@ query_time_seconds(void)
 {
     /* kernel_query_time_seconds() returns seconds since the Unix Epoch
      * (1970-01-01), but DR's cross-platform contract for query_time_seconds()
-     * is seconds since 1601-01-01 (see os_shared.h). Convert so the kernel
-     * build matches the user space build.
+     * is seconds since 1601-01-01 (see os_shared.h). add UTC_TO_EPOCH_SECONDS
+     * so the kernel build matches the user space build.
      */
-    return (uint)(kernel_query_time_seconds() + UTC_TO_EPOCH_SECONDS);
+    return (kernel_query_time_seconds() + UTC_TO_EPOCH_SECONDS);
 }
